refactor: Improve Maintenance entity integrity and structure.
This refactor tries to apply some Rich Domain structure into the entity.
Updated equals() and hashcode() to handle UUID generated at the object
creation to fix the skill issue from before. Was handling ID which was
set by the database, it was null on object creation thus breaking the
contract of the interfaces that depends on hashcode.
Entity gets the knowledge to update its own updatable fields.
Encapsulated state transition rules and update logic within the entity,
trying to keep service layer simpler.