feat: Implement endpoint for registering new Pledge for a Maintenance record.
Implement the normal flow of new records. RequestBody mapped to a dto
into the service layer that makes the mapping and persist the entity.
Implements a new method for the `MaintenanceService` to return the
entity using its publicId for associations that demands the 'internal'
id field.
Implements a `PledgeMapper` to standardize ins/outs of object mapping.

Closes #11