feat: Implement endpoint to register MaintenanceAction.
Implements the normal flow for inserting new records into the database.
`Maintenance` entity contains new method to check for completion to
comply with business rules.
Adds `ActionMaterial` entity that encapsulates a material used.
This `ActionMaterial` should not exists by itself and should always be
handled via its parent `MaintenanceAction`.
This implementation still lacks in desision making, if any
`MaintenanceAction` should be "atomic" in a sense of existing
perpetually and never be updated to mantain consistency.

Refs #14